Moved mailing components to its own app
