+++ backend/ambassadors/models.py
@@ -202,55 +202,3 @@ class AmbassadorSize(models.Model):
         return (
             f"Размер обуви:{self.foot_size}, размер одежды:{self.clothes_size}"
         )
-
-
-class Message(models.Model):
-    """
-    Модель отправки сообщений.
-    """
-
-    ambassadors = models.ManyToManyField(Ambassador, related_name="messages")
-    title = models.CharField(verbose_name="Заголовок", max_length=100)
-    text = models.CharField(verbose_name="Текст", max_length=2000)
-    sent = models.DateTimeField(
-        verbose_name="Дата отправки", null=True, default=timezone.now
-    )
-    by_email = models.BooleanField(verbose_name="На почту", default=False)
-    to_telegram = models.BooleanField(verbose_name="В телеграм", default=False)
-    is_sent = models.BooleanField(verbose_name="Отправить", default=False)
-
-    class Meta:
-        verbose_name = "Сообщение для амбассадора"
-        verbose_name_plural = "Сообщения для амбассадоров"
-        ordering = ("id",)
-
-    def __str__(self) -> str:
-        return f"{self.title}"
-
-
-# class MessageToAmbassador(models.Model):
-#     """
-#     Связанная таблица сообщений и амбассадоров.
-#     """
-
-#     ambassador = models.ForeignKey(
-#         Ambassador,
-#         verbose_name="Амбассадор",
-#         related_name="messages",
-#         on_delete=models.SET_NULL,
-#         null=True,
-#     )
-#     message = models.ForeignKey(
-#         Message,
-#         verbose_name="Сообщение",
-#         on_delete=models.SET_NULL,
-#         null=True,
-#     )
-
-#     class Meta:
-#         verbose_name = "Отправленное сообщение для амбассадора"
-#         verbose_name_plural = "Отправленные сообщения для амбассадоров"
-#         ordering = ("id",)
-
-#     def __str__(self) -> str:
-#         return f"{self.ambassador} {self.message}"
