+++ backend/ambassadors/serializers.py
@@ -1,18 +1,14 @@
-from django.db import transaction
 from rest_framework import serializers
-from rest_framework.exceptions import ValidationError
 
-from content.models import Content, Promocode
-
-from .models import (
+from ambassadors.models import (
     Actions,
     Ambassador,
     AmbassadorActions,
     AmbassadorAddress,
     AmbassadorSize,
-    Message,
     YandexProgramm,
 )
+from content.models import Content, Promocode
 
 
 class YandexProgrammSerializer(serializers.ModelSerializer):
@@ -290,56 +286,3 @@ class AmbassadorInMessageSerializer(serializers.ModelSerializer):
     class Meta:
         model = Ambassador
         fields = ("id", "tg_acc")
-
-
-class MessageSerializer(serializers.ModelSerializer):
-    """Сериализатор для модели сообщений."""
-
-    ambassadors = AmbassadorInMessageSerializer(many=True)
-    sent = serializers.DateTimeField(required=False, allow_null=True)
-    by_email = serializers.BooleanField(required=False)
-    to_telegram = serializers.BooleanField(required=False)
-    is_sent = serializers.BooleanField(required=False)
-
-    class Meta:
-        model = Message
-        fields = (
-            "id",
-            "title",
-            "text",
-            "sent",
-            "by_email",
-            "to_telegram",
-            "is_sent",
-            "ambassadors",
-        )
-
-    def create(self, validated_data):
-        ambassadors_data = validated_data.pop("ambassadors", [])
-        with transaction.atomic():
-            instance = Message.objects.create(**validated_data)
-            ambassadors_id = [amb["id"] for amb in ambassadors_data]
-            ambassadors = Ambassador.objects.filter(id__in=(ambassadors_id))
-            if len(ambassadors) != len(ambassadors_id):
-                raise ValidationError(
-                    "Абмассадора с таким id не существует", 404
-                )
-            instance.ambassadors.set(ambassadors_id)
-        return instance
-
-    def update(self, instance: Message, validated_data):
-        ambassadors_data = validated_data.pop("ambassadors", [])
-        if validated_data.get("is_sent") is True:
-            raise ValidationError("Это сообщение уже нельзя изменить.")
-        with transaction.atomic():
-            for attr, value in validated_data.items():
-                setattr(instance, attr, value)
-            ambassadors_id = [amb["id"] for amb in ambassadors_data]
-            ambassadors = Ambassador.objects.filter(id__in=(ambassadors_id))
-            if len(ambassadors) != len(ambassadors_id):
-                raise ValidationError(
-                    "Абмассадора с таким id не существует", 404
-                )
-            instance.ambassadors.set(ambassadors_id)
-            instance.save()
-        return instance
