+++ backend/ambassadors_project/settings.py
@@ -37,6 +37,7 @@ INSTALLED_APPS = [
     "content",
     "bot",
     "websocket",
+    "mailing",
 ]
 
 MIDDLEWARE = [
