+++ docker-compose.production.yml
@@ -25,14 +25,12 @@ services:
       - 6379
 
   worker:
-    restart: always
     build: ./backend/
     command: celery -A ambassadors_project worker -B -l info
     env_file: .env
     volumes:
       - static:/backend_static
       - media:/app/media
-      - ./backend/:/app/
     depends_on:
       - backend
       - redis
